refactor: move testutils to internal
