+++ internal/centralizedmetrics/providers/mixpanel_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/status-im/status-go/internal/centralizedmetrics/common"
-	"github.com/status-im/status-go/pkg/testutils"
+	"github.com/status-im/status-go/internal/testutils"
 )
 
 func TestMixpanelMetricProcessor(t *testing.T) {
