+++ messaging/layers/encryption/publisher/publisher_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
 
-	"github.com/status-im/status-go/pkg/testutils"
+	"github.com/status-im/status-go/internal/testutils"
 )
 
 func TestServiceTestSuite(t *testing.T) {
