+++ messaging/layers/transport/filters_manager_test.go
@@ -10,12 +10,12 @@ import (
 
 	"github.com/libp2p/go-libp2p/core/peer"
 
-	wakuv2 "github.com/status-im/status-go/messaging/waku"
-	"github.com/status-im/status-go/pkg/testutils"
-
 	_ "github.com/mutecomm/go-sqlcipher/v4"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
+
+	"github.com/status-im/status-go/internal/testutils"
+	wakuv2 "github.com/status-im/status-go/messaging/waku"
 
 	"github.com/status-im/status-go/crypto"
 )
