+++ protocol/messenger_identity_display_name_test.go
@@ -5,12 +5,12 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/stretchr/testify/suite"
+
 	utils "github.com/status-im/status-go/common"
 	"github.com/status-im/status-go/internal/db/multiaccounts/accounts"
+	"github.com/status-im/status-go/internal/testutils"
 	messagingtypes "github.com/status-im/status-go/messaging/types"
-	"github.com/status-im/status-go/pkg/testutils"
-
-	"github.com/stretchr/testify/suite"
 )
 
 const testDisplayName = "My New Display Name"
