+++ protocol/messenger_sync_settings_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/status-im/status-go/crypto/types"
 	settings2 "github.com/status-im/status-go/internal/db/multiaccounts/settings"
+	"github.com/status-im/status-go/internal/testutils"
 	messagingtypes "github.com/status-im/status-go/messaging/types"
-	"github.com/status-im/status-go/pkg/testutils"
 	"github.com/status-im/status-go/services/stickers"
 )
