+++ protocol/push_notification_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/status-im/status-go/crypto"
 	"github.com/status-im/status-go/crypto/types"
-	"github.com/status-im/status-go/pkg/testutils"
+	testutils2 "github.com/status-im/status-go/internal/testutils"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/communities"
 	"github.com/status-im/status-go/protocol/contacts"
@@ -93,7 +93,7 @@ func (s *MessengerPushNotificationSuite) newPushNotificationServer() (*Messenger
 
 	serverConfig := &pushnotificationserver.Config{
 		Enabled:   true,
-		Logger:    testutils.MustCreateTestLogger(),
+		Logger:    testutils2.MustCreateTestLogger(),
 		Identity:  privateKey,
 		GorushURL: s.gorushMockURL,
 	}
@@ -131,7 +131,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotification() {
 	err = bob1.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -170,7 +170,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotification() {
 	err = bob2.RegisterForPushNotifications(context.Background(), bob2DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 	s.Require().NoError(err)
 
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -215,7 +215,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotification() {
 	s.Require().NoError(err)
 
 	infoMap := make(map[string]*pushnotificationclient.PushNotificationInfo)
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -266,7 +266,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotification() {
 	s.Require().Len(retrievedNotificationInfo, 2)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -324,7 +324,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationFromContactO
 	s.Require().NoError(err)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -368,7 +368,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationFromContactO
 	s.Require().NoError(err)
 
 	var info []*pushnotificationclient.PushNotificationInfo
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -402,7 +402,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationFromContactO
 	s.Require().Len(retrievedNotificationInfo, 1)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -475,7 +475,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationRetries() {
 	s.Require().NoError(err)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -517,7 +517,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationRetries() {
 
 	// We check that alice retrieves the info from the messenger
 	var info []*pushnotificationclient.PushNotificationInfo
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -555,7 +555,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationRetries() {
 
 	// Re-registration should be triggered, pull from messenger and bob to check we are correctly registered
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -587,7 +587,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationRetries() {
 	messageID, err := hex.DecodeString(messageIDString[2:])
 	s.Require().NoError(err)
 
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -624,7 +624,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationRetries() {
 	s.Require().Len(retrievedNotificationInfo, 1)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -668,7 +668,7 @@ func (s *MessengerPushNotificationSuite) TestContactCode() {
 	err = bob1.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -740,7 +740,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationMention() {
 	err = bob.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -783,7 +783,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationMention() {
 	s.Require().NoError(err)
 
 	var bobInfo []*pushnotificationclient.PushNotificationInfo
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -819,7 +819,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationMention() {
 	s.Require().Len(retrievedNotificationInfo, 1)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -862,7 +862,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationCommunityReq
 	err = bob.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -922,7 +922,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationCommunityReq
 	s.NoError(err)
 
 	// Pull message and make sure org is received
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		response, err = alice.RetrieveAll()
 		if err != nil {
 			return err
@@ -950,7 +950,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationCommunityReq
 	s.Require().Equal(requestToJoin1.PublicKey, crypto.PubkeyToHex(&alice.identity.PublicKey))
 	s.Require().Equal(communities.RequestToJoinStatePending, requestToJoin1.State)
 
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -995,7 +995,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationPairedDevice
 	err = bob1.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1034,7 +1034,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationPairedDevice
 	err = bob2.RegisterForPushNotifications(context.Background(), bob2DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 	s.Require().NoError(err)
 
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1079,7 +1079,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationPairedDevice
 	s.Require().NoError(err)
 
 	infoMap := make(map[string]*pushnotificationclient.PushNotificationInfo)
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1130,7 +1130,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationPairedDevice
 	s.Require().Len(retrievedNotificationInfo, 2)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1188,7 +1188,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationReply() {
 	err = bob.RegisterForPushNotifications(context.Background(), bob1DeviceToken, testAPNTopic, protobuf.PushNotificationRegistration_APN_TOKEN)
 
 	// Pull servers  and check we registered
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1251,7 +1251,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationReply() {
 	s.Require().NoError(err)
 
 	var bobInfo []*pushnotificationclient.PushNotificationInfo
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
@@ -1287,7 +1287,7 @@ func (s *MessengerPushNotificationSuite) TestReceivePushNotificationReply() {
 	s.Require().Len(retrievedNotificationInfo, 1)
 
 	var sentNotification *pushnotificationclient.SentNotification
-	err = testutils.RetryWithBackOff(func() error {
+	err = testutils2.RetryWithBackOff(func() error {
 		_, err = messenger.RetrieveAll()
 		if err != nil {
 			return err
