+++ services/newsfeed/service_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/mock/gomock"
 
-	"github.com/status-im/status-go/pkg/testutils"
+	"github.com/status-im/status-go/internal/testutils"
 	"github.com/status-im/status-go/protocol"
 	"github.com/status-im/status-go/services/newsfeed/migrations"
 	mock_newsfeed "github.com/status-im/status-go/services/newsfeed/mock"
